refactor(swap): extract script/verify/derive helpers out of rfq.ts
rfq.ts should contain only transport and comms logic. Moves:
- VTXO script builders to vhtlcScript.ts
- Error classes to rfqErrors.ts
- Verification/assertion helpers to rfqVerify.ts
- Corridor derivation functions to rfqDerive.ts

No exported API changes.
