+++ packages/swap/src/index.ts
@@ -93,43 +93,48 @@ export {
     LIGHTNING_BTC,
     LIGHTNING_RECEIVE_PAIR,
     LIGHTNING_SEND_PAIR,
-    MIN_CLAIM_WINDOW_SECONDS,
-    MIN_HEADROOM_SECONDS,
-    RFQ_TERMINAL_STATES,
-    SOLO_REFUND_HEADROOM_SECONDS,
-    AddressMismatch,
-    SwapRefusal,
     arkadeAssetLeg,
     arkadeSwapRequest,
-    assertFundable,
-    assertReceivable,
-    deriveLightningReceive,
-    deriveOnchainReceive,
     httpTransport,
     lightningReceiveRequest,
     lightningSendRequest,
-    lightningSendVtxoScript,
     newRfqId,
     offerTermsFromQuote,
-    receiveVtxoScript,
     relayTransport,
     requestLightningReceive,
     requestLightningSend,
     rfqPair,
-    unilateralClaimDelay,
-    unilateralRefundDelay,
-    unilateralRefundWithoutReceiverDelay,
-    verifyLockupAddress,
-    verifyReceiveInvoice,
-    type InvoiceFacts,
-    type LightningReceiveTreeParams,
-    type LightningSendTreeParams,
     type RelaySocket,
     type RfqQuote,
-    type RfqRefusalReason,
     type RfqStatus,
     type RfqTransport,
 } from "./rfq";
+export {
+    MIN_CLAIM_WINDOW_SECONDS,
+    MIN_HEADROOM_SECONDS,
+    assertFundable,
+    assertReceivable,
+    verifyLockupAddress,
+    verifyReceiveInvoice,
+    type InvoiceFacts,
+} from "./rfqVerify";
+export {
+    RFQ_TERMINAL_STATES,
+    AddressMismatch,
+    SwapRefusal,
+    type RfqRefusalReason,
+} from "./rfqErrors";
+export {
+    SOLO_REFUND_HEADROOM_SECONDS,
+    lightningSendVtxoScript,
+    receiveVtxoScript,
+    unilateralClaimDelay,
+    unilateralRefundDelay,
+    unilateralRefundWithoutReceiverDelay,
+    type LightningReceiveTreeParams,
+    type LightningSendTreeParams,
+} from "./vhtlcScript";
+export { deriveLightningReceive, deriveOnchainReceive } from "./rfqDerive";
 export {
     LOCKTIME_THRESHOLD,
     ONCHAIN_DUST_SATS,
@@ -158,12 +163,12 @@ export {
     MAX_MIN_CONFIRMATIONS,
     ONCHAIN_CLAIM_MARGIN_SECONDS,
     ONCHAIN_ORDER_MARGIN_SECONDS,
-    deriveOnchainSend,
     onchainReceiveRequest,
     onchainSendRequest,
     requestOnchainReceive,
     requestOnchainSend,
 } from "./rfq";
+export { deriveOnchainSend } from "./rfqDerive";
 export { sealClaimPacket, type ClaimPacketInput, type SealedClaimPacket } from "./claimPacket";
 export {
     LockupAmountMismatchError,
