+++ packages/swap/src/rfqErrors.ts
@@ -1,0 +1,44 @@
+/**
+ * Error classes and closed sets for RFQ swap corridors.
+ *
+ * Extracted from rfq.ts — error types only, no transport, no verification
+ * logic. Imported by rfqVerify.ts and rfq.ts.
+ */
+
+/** The closed refusal set. Treat any unknown reason as a generic decline. */
+export type RfqRefusalReason =
+    | "unsupported_pair"
+    | "unsupported_payload"
+    | "amount_out_of_range"
+    | "exposure_cap"
+    | "invoice_expired"
+    | "quote_conflict"
+    | "pricing_unavailable"
+    | "rate_limited";
+
+/** Lifecycle vocabulary; states after which nothing more will happen. */
+export const RFQ_TERMINAL_STATES = ["settled", "refused", "expired", "refunded", "stuck"] as const;
+
+/** A refusal from the solver, carrying its closed-set reason. */
+export class SwapRefusal extends Error {
+    readonly reason: string;
+    readonly rfqId: string | undefined;
+    constructor(reason: string, rfqId?: string) {
+        super(`solver refused: ${reason}`);
+        this.name = "SwapRefusal";
+        this.reason = reason;
+        this.rfqId = rfqId;
+    }
+}
+
+/** The solver's address does not match the local derivation. NEVER fund past this. */
+export class AddressMismatch extends Error {
+    readonly derived: string;
+    readonly quoted: string | undefined;
+    constructor(derived: string, quoted?: string) {
+        super("solver lockup address does not match local derivation — refusing to fund");
+        this.name = "AddressMismatch";
+        this.derived = derived;
+        this.quoted = quoted;
+    }
+}
